% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Small_misc_functions.R
\name{custom_round}
\alias{custom_round}
\title{Round up any numbers with .5 - R will default round to an even number with the round function.}
\usage{
custom_round(data, digits = 0)
}
\arguments{
\item{data}{The column with data to be rounded}
}
\value{
Data rounded up if at .5 or above and down if below .5
}
\description{
Round up any numbers with .5 - R will default round to an even number with the round function.
}
\examples{
custom_round((BUILDING_HEIGHT/3), 0)
}
