% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Address_matcher.R
\name{epc_address_matching}
\alias{epc_address_matching}
\title{Matching EPC addresses to Ordnance Survey UPRNs}
\usage{
epc_address_matching(
  addressesToMatch,
  rowID = BUILDING_REFERENCE_NUMBER,
  add1 = ADDRESS1,
  add2 = ADDRESS2,
  add3 = ADDRESS,
  pcode = POSTCODE,
  country = "S",
  sql = R_ADDRBASE_PLUS_20231014,
  path = NULL
)
}
\arguments{
\item{addressesToMatch}{The dataframe with the addresses to be matched}

\item{rowID}{column with unique property key. Important to be able to join address matching results back to original EPC data.}

\item{add1}{column with first line of address details. Intended to be building/property identifier (name/number), but often includes street name}

\item{add2}{column with second line of address details. Intended to be street name, but often includes building identifiers}

\item{add3}{column with third line of address details. Intended to be town/city}

\item{pcode}{column with post code of address. Full postcode (e.g. EH21 6TU). Key for filtering down relevant addresses to match against.}

\item{country}{optional column to set the country addresses are from. Options = E, S, W. Multiple allowed, but must be in quotations.}

\item{sql}{sql server table to pull AddressBase data from. Requires a valid ODBC connection.}

\item{path}{location to save interim and final files. Provide in quotations. Default is NULL and these will save to a generic location.}
}
\value{
A dataframe with the UPRN for each address
}
\description{
Standarised matching procedure for Home Analytics updates
}
\examples{
epc_address_matching(addressesToMatch)
epc_address_matching(addressesToMatch, rowID, add1 = ADD_1, add2 = ADD_2, add3 = ADD_3, pcode = POSTIE, country = "S", sql = R_ADDRBASE_PLUS_20231014, path = "C:/Users/Data/")
epc_address_matching(addressesToMatch, rowID, add1 = ADD_1, add2 = ADD_2, add3 = ADD_3, pcode = POSTIE, country = c("E", "W"), sql = R_ADDRBASE_PLUS_20231014, path = "C:/Users/Data/")
}
