% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_functions.R
\name{epc_clean_commerical}
\alias{epc_clean_commerical}
\title{EPC cleaning 'master' wrapper for Scotland EPCs}
\usage{
epc_clean_commerical(data, func_list = NULL, mute = NULL)
}
\arguments{
\item{data}{The dataframe with EPC data to be cleaned}

\item{func_list}{a list of the functions to be run (optional)}

\item{mute}{an option to toggle print statements for each function. Defaults to NULL and will always print all tables and statements. Set to TRUE to mute these.}
}
\value{
Dataframes with UPRN, cleaned data and original EPC data for onward processing
}
\description{
Very simple wrapper that runs through each EPC cleaning function for the commercial Scotland HA update.
}
\examples{
epc_clean_commerical()
epc_clean_commerical(EPC_DOWNLOAD)
}
