% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neighbhourhood_measures_continuous.R
\name{neigh_meas_cont}
\alias{neigh_meas_cont}
\title{Wrapper for building out neighbhourhood measures for Home Analytics}
\usage{
neigh_meas_cont(df, var, geogs = NULL, ptype = PTYPE_SUBGROUP)
}
\arguments{
\item{df}{The dataframe with EPC data to be cleaned}

\item{var}{the variable column to be used with continuous, numeric data}

\item{geogs}{an option to provide a list of geographies to compute neighbhourhood measure's proportion/uncertainty for}

\item{ptype}{the property type subgroup to compute neighbhourhood measures against. Expects column called PTYPE_SUBGROUP, but can provide your own.}
}
\value{
Dataframes with UPRN, cleaned data and original EPC data for onward processing
}
\description{
Very simple wrapper that runs through each EPC cleaning function for the commercial Scotland HA update. Works for building he continuous variables
}
\examples{
neigh_meas(EPCs, TOTAL_FLOOR_ARE, list, PROPERTY_TYPE_GROUPS)
}
