% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPC_Cleaning_SComm.R
\name{SComm_age}
\alias{SComm_age}
\title{Property age cleaning for England & Wales data}
\usage{
SComm_age(
  data,
  age = CONSTRUCTION_AGE_BAND,
  epc = CURRENT_ENERGY_RATING,
  mute = NULL,
  ...
)
}
\arguments{
\item{data}{The dataframe with data to be cleaned}

\item{age}{The property age column in the dataframe to be cleaned}

\item{mute}{option to mute some print statements. If set to TRUE then only prints when processing starts and ends. Default is NULL, but FALSE is also valid.}
}
\value{
A dataframe with the UPRN, cleaned data and original EPC data for onward processing
}
\description{
Standarised cleaning procedure for Home Analytics updates
}
\examples{
SComm_age(data)
SComm_age(data, AGE_OF_PROPERTY, BANDS)
}
