% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPC_Cleaning_SComm.R
\name{SComm_bills}
\alias{SComm_bills}
\title{SAP derived variables from Scotland EPC data}
\usage{
SComm_bills(
  data,
  heat_cost = HEATING_COST_CURRENT,
  water_cost = HOT_WATER_COST_CURRENT,
  lighting_cost = LIGHTING_COST_CURRENT,
  mute = NULL,
  ...
)
}
\arguments{
\item{data}{The dataframe with data to be cleaned}

\item{heat_cost}{The column in the dataframe containing space heating cost}

\item{water_cost}{The column in the dataframe containing water heating cost}

\item{lighting_cost}{The column in the dataframe containing lighting cost}

\item{mute}{option to mute some print statements. If set to TRUE then only prints when processing starts and ends. Default is NULL, but FALSE is also valid.}
}
\value{
A dataframe with the UPRN, cleaned data and original EPC data for onward processing
}
\description{
Standarised cleaning procedure for Home Analytics updates
}
\examples{
SComm_bills(data)
SComm_bills(data, HEAT_COST, WATER_HEAT_COST, LIGHTING_COST)
}
