% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPC_Cleaning_SComm.R
\name{scomm_demand}
\alias{scomm_demand}
\title{SAP derived variables from Scotland EPC data}
\usage{
scomm_demand(
  data,
  space_heat = SPACE_HEATING_DEMAND,
  water_heat = WATER_HEATING_DEMAND,
  mute = NULL,
  ...
)
}
\arguments{
\item{data}{The dataframe with data to be cleaned}

\item{space_heat}{The column in the dataframe containing space heating demand}

\item{water_heat}{The column in the dataframe containing water heating demand}

\item{mute}{option to mute some print statements. If set to TRUE then only prints when processing starts and ends. Default is NULL, but FALSE is also valid.}
}
\value{
A dataframe with the UPRN, cleaned data and original EPC data for onward processing
}
\description{
Standarised cleaning procedure for Home Analytics updates
}
\examples{
scomm_demand(data)
scomm_demand(data, SAP_SCORE, ENERGY_USAGE, FLOOR_AREA)
}
