% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPC_Cleaning_SComm.R
\name{scomm_energy}
\alias{scomm_energy}
\title{SAP derived variables from Scotland EPC data}
\usage{
scomm_energy(
  data,
  sap = CURRENT_ENERGY_EFFICIENCY,
  energy = ENERGY_CONSUMPTION_CURRENT,
  tfa = TOTAL_FLOOR_AREA,
  mute = NULL,
  ...
)
}
\arguments{
\item{data}{The dataframe with data to be cleaned}

\item{sap}{The column in the dataframe containing SAP score}

\item{energy}{The column in the dataframe containing energy usage (in this case kwh/m2/year).}

\item{tfa}{The column in the dataframe containing total floor area (to convert energy/m2 into total energy).}

\item{mute}{option to mute some print statements. If set to TRUE then only prints when processing starts and ends. Default is NULL, but FALSE is also valid.}
}
\value{
A dataframe with the UPRN, cleaned data and original EPC data for onward processing
}
\description{
Standarised cleaning procedure for Home Analytics updates
}
\examples{
scomm_energy(data)
scomm_energy(data, SAP_SCORE, ENERGY_USAGE, FLOOR_AREA)
}
