% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPC_Cleaning_SComm.R
\name{scomm_energy_gen}
\alias{scomm_energy_gen}
\title{Energy generation data cleaning for Scotland EPC data}
\usage{
scomm_energy_gen(
  data,
  lzc = LZC_ENERGY_SOURCES,
  pv = PHOTO_SUPPLY,
  therm = SOLAR_WATER_HEATING_FLAG,
  wind = WIND_TURBINE_COUNT,
  mute = NULL,
  ...
)
}
\arguments{
\item{data}{The dataframe with data to be cleaned}

\item{lzc}{The column in the dataframe containing low zero carbon (LZC) tech}

\item{pv}{The column in the dataframe containing PV data. Expects numbers to be extractable after certain strings (e.g. ... Peak Power: 3.5 ...)}

\item{therm}{The column in the dataframe containing solar thermal flags (true, Y, N, false).}

\item{wind}{The column in the dataframe containing wind turbine counts.}

\item{mute}{option to mute some print statements. If set to TRUE then only prints when processing starts and ends. Default is NULL, but FALSE is also valid.}
}
\value{
A dataframe with the UPRN, cleaned data and original EPC data for onward processing
}
\description{
Standarised cleaning procedure for Home Analytics updates
}
\examples{
scomm_energy_gen(data)
scomm_energy_gen(data, LZC_ENERGY, PV_FLAG, THERMAL_FLAG, WIND_FLAG)
}
