% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPC_Cleaning_SComm.R
\name{scomm_environmental}
\alias{scomm_environmental}
\title{SAP derived variables from Scotland EPC data}
\usage{
scomm_environmental(
  data,
  carbon = CO2_EMISSIONS_CURRENT,
  env = ENVIRONMENT_IMPACT_CURRENT,
  mute = NULL,
  ...
)
}
\arguments{
\item{data}{The dataframe with data to be cleaned}

\item{carbon}{The column in the dataframe containing current carbon emissions estimate}

\item{env}{The column in the dataframe containing environmental impact score}

\item{mute}{option to mute some print statements. If set to TRUE then only prints when processing starts and ends. Default is NULL, but FALSE is also valid.}
}
\value{
A dataframe with the UPRN, cleaned data and original EPC data for onward processing
}
\description{
Standarised cleaning procedure for Home Analytics updates
}
\examples{
scomm_environmental(data)
scomm_environmental(data, CO2_CURRENT, WATER_HEAT_COST, LIGHTING_COST)
}
