% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPC_Cleaning_SComm.R
\name{SComm_flatfloor}
\alias{SComm_flatfloor}
\title{Flat floor level cleaning for Scotland EPC data}
\usage{
SComm_flatfloor(
  data,
  floor = FLOOR_LEVEL,
  ptype = PROPERTY_TYPE,
  mute = NULL,
  ...
)
}
\arguments{
\item{data}{The dataframe with data to be cleaned}

\item{floor}{The column in the dataframe to be cleaned with flat floor info.}

\item{ptype}{The column in the dataframe with property type data.}

\item{mute}{option to mute some print statements. If set to TRUE then only prints when processing starts and ends. Default is NULL, but FALSE is also valid.}
}
\value{
A dataframe with the UPRN, cleaned data and original EPC data for onward processing
}
\description{
Standarised cleaning procedure for Home Analytics updates
}
\examples{
SComm_flatfloor(data)
SComm_flatfloor(data, FLOORS, PTYPE)
}
