% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPC_Cleaning_SComm.R
\name{scomm_floor}
\alias{scomm_floor}
\title{Floor Construction and insulation cleaning for Scotland EPCs}
\usage{
scomm_floor(
  data,
  floors = FLOOR_DESCRIPTION,
  ages = CONSTRUCTION_AGE_BAND,
  imp = IMPROVEMENTS,
  mute = NULL,
  ...
)
}
\arguments{
\item{data}{The dataframe with data to be cleaned}

\item{floors}{The column in the dataframe to be cleaned.}

\item{ages}{The property age column in the dataframe to help infer insulation thickness.}

\item{imp}{The column in the dataframe with retrofit improvements. Expecting a column with long strings with measures seperated by a pipe (|) delimiter}

\item{mute}{option to mute some print statements. If set to TRUE then only prints when processing starts and ends. Default is NULL, but FALSE is also valid.}
}
\value{
A dataframe with the UPRN, cleaned data and original EPC data for onward processing
}
\description{
Standarised cleaning procedure for Home Analytics updates
}
\examples{
scomm_floor(data)
scomm_floor(data, FLOOR_TYPE, AGE_BANDS)
}
