% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPC_Cleaning_SComm.R
\name{scomm_gas_grid}
\alias{scomm_gas_grid}
\title{Gas grid inference from Scotland EPCs}
\usage{
scomm_gas_grid(
  data,
  grid = MAINS_GAS_FLAG,
  m_heat = MAINHEAT_DESCRIPTION,
  s_heat = SECONDHEAT_DESCRIPTION,
  mute = NULL,
  ...
)
}
\arguments{
\item{data}{The dataframe with data to be cleaned}

\item{grid}{The column in the dataframe with gas grid flag (expects "Y" and "N")}

\item{m_heat}{The column in the dataframe with main heating descriptions (if mains gas is used somewhere then assumed grid connection)}

\item{s_heat}{The column in the dataframe with secondary heating descriptions(if mains gas is used somewhere then assumed grid connection)}

\item{mute}{option to mute some print statements. If set to TRUE then only prints when processing starts and ends. Default is NULL, but FALSE is also valid.}
}
\value{
A dataframe with the UPRN, cleaned data and original EPC data for onward processing
}
\description{
Standarised cleaning procedure for Home Analytics updates
}
\examples{
scomm_gas_grid(data)
scomm_gas_grid(data, GAS_GRID_FLAG, MAIN_HEATING_SYSTEM_DETAILS, SEC_HEATING_SYSTEM_DETAILS)
}
