% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPC_Cleaning_SComm.R
\name{SComm_glazing}
\alias{SComm_glazing}
\title{Glazing data cleaning for Scotland EPC data}
\usage{
SComm_glazing(
  data,
  glaz = WINDOWS_DESCRIPTION,
  type = GLAZED_TYPE,
  multi = MULTI_GLAZE_PROPORTION,
  eff = WINDOWS_ENERGY_EFF,
  age = CONSTRUCTION_AGE_BAND,
  imp = IMPROVEMENTS,
  mute = NULL,
  ...
)
}
\arguments{
\item{data}{The dataframe with data to be cleaned}

\item{glaz}{The column in the dataframe to be cleaned.}

\item{multi}{The column in the dataframe to be cleaned with multiple glazing types.}

\item{age}{The column in the dataframe containing property age data.}

\item{imp}{The column in the dataframe with retrofit improvements. Expecting a column with long strings with measures seperated by a pipe (|) delimiter}

\item{mute}{option to mute some print statements. If set to TRUE then only prints when processing starts and ends. Default is NULL, but FALSE is also valid.}
}
\value{
A dataframe with the UPRN, cleaned data and original EPC data for onward processing
}
\description{
Standarised cleaning procedure for Home Analytics updates
}
\examples{
SComm_glazing(data)
SComm_glazing(data, GLAZING, type, multi)
}
