% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPC_Cleaning_SComm.R
\name{scomm_rooms}
\alias{scomm_rooms}
\title{Habitable rooms and total floor area cleaning for Scotland EPCs}
\usage{
scomm_rooms(
  data,
  rooms = NUMBER_HABITABLE_ROOMS,
  tfa = TOTAL_FLOOR_AREA,
  mute = NULL,
  ...
)
}
\arguments{
\item{data}{The dataframe with data to be cleaned}

\item{rooms}{The column in the dataframe to be cleaned with habitable room counts. Expects number strings (e.g. "7" not "seven")}

\item{tfa}{The column in the dataframe containing floor area.}

\item{mute}{option to mute some print statements. If set to TRUE then only prints when processing starts and ends. Default is NULL, but FALSE is also valid.}
}
\value{
A dataframe with the UPRN, cleaned data and original EPC data for onward processing
}
\description{
Standarised cleaning procedure for Home Analytics updates
}
\examples{
scomm_rooms(data)
scomm_rooms(data, HABITABLE_ROOMS_NEW, TFA)
}
