% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPC_Cleaning_SComm.R
\name{SComm_sec_heating}
\alias{SComm_sec_heating}
\title{Secondary Heating Fuel and System cleaning for Scotland EPCs}
\usage{
SComm_sec_heating(data, s_heat = SECONDHEAT_DESCRIPTION, mute = NULL, ...)
}
\arguments{
\item{data}{The dataframe with data to be cleaned}

\item{s_heat}{The column in the dataframe to be cleaned}

\item{mute}{option to mute some print statements. If set to TRUE then only prints when processing starts and ends. Default is NULL, but FALSE is also valid.}
}
\value{
A dataframe with the UPRN, cleaned data and original EPC data for onward processing
}
\description{
Standarised cleaning procedure for Home Analytics updates
}
\examples{
SComm_sec_heating(data)
SComm_sec_heating(data, SEC_HEATING_SYSTEM_DETAILS)
}
