% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPC_Cleaning_SComm.R
\name{scomm_type}
\alias{scomm_type}
\title{Property type cleaning for Scotland EPCs}
\usage{
scomm_type(data, type = PROPERTY_TYPE, form = BUILT_FORM, mute = NULL, ...)
}
\arguments{
\item{data}{The dataframe with data to be cleaned}

\item{type}{The property type column in the dataframe}

\item{form}{The property form column in the dataframe}

\item{mute}{option to mute some print statements. If set to TRUE then only prints when processing starts and ends. Default is NULL, but FALSE is also valid.}
}
\value{
A dataframe with the UPRN, cleaned data and original EPC data for onward processing
}
\description{
Standarised cleaning procedure for Home Analytics updates
}
\examples{
scomm_type(data)
scomm_type(data, type = PTYPE_NEW, form = B_FORM, mute = TRUE)
}
