% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOID_BB_Logic.R
\name{TOID_BB_logic}
\alias{TOID_BB_logic}
\title{TOID and Building Block inference logic}
\usage{
TOID_BB_logic(data, var, group, trust = 0.5, join = TRUE)
}
\arguments{
\item{data}{the dataframe with data to be processed.}

\item{var}{the column with the property characteristics of interest.}

\item{group}{the spatial unit - either TOID or Building Block.}

\item{trust}{the proportion of which to "trust" that a certain characteristics is the majority of known records. Default is 0.5 - i.e. needs to be above 50 percent of one label to be classed as the trusted label for that spatial unit.}

\item{join}{option to join the results back to the original input dataset. Set to FALSE to keep the output as a stand alone data table.}
}
\value{
a dataframe with the UPRN and trusted label for property characteristics of interest or the original dataset with a column containing the trusted label.
}
\description{
Function to infer more "trusted" property characteristics based on proportions of characteristics in other properties within the TOID/Building Block
}
\examples{
TOID_BB_logic(dataset, FINAL_PROPERTY_AGE, BB_ID)
TOID_BB_logic(dataset, FINAL_PROPERTY_AGE, OS_TOPO_TOID, trust = 0.75, join = FALSE)
}
